+++ deployments/arbitrumGoerli/SignatureLib.json
@@ -1,0 +1,82 @@
+{
+  "address": "0xA8Fc4232581deC73e65C6095d410b480c6EaB56E",
+  "abi": [
+    {
+      "inputs": [
+        {
+          "internalType": "bytes32",
+          "name": "message",
+          "type": "bytes32"
+        },
+        {
+          "components": [
+            {
+              "internalType": "uint8",
+              "name": "v",
+              "type": "uint8"
+            },
+            {
+              "internalType": "bytes32",
+              "name": "r",
+              "type": "bytes32"
+            },
+            {
+              "internalType": "bytes32",
+              "name": "s",
+              "type": "bytes32"
+            }
+          ],
+          "internalType": "struct SignatureLib.Signature",
+          "name": "signature",
+          "type": "tuple"
+        }
+      ],
+      "name": "recoverSigner",
+      "outputs": [
+        {
+          "internalType": "address",
+          "name": "",
+          "type": "address"
+        }
+      ],
+      "stateMutability": "pure",
+      "type": "function"
+    }
+  ],
+  "transactionHash": "0x3572a3eef62da1b223a67567e89066dcd52b3857c7e34c188367f5701c0d08a3",
+  "receipt": {
+    "to": null,
+    "from": "0x1Ab4C6d9e25Fc65C917aFBEfB4E963C400Fb9814",
+    "contractAddress": "0xA8Fc4232581deC73e65C6095d410b480c6EaB56E",
+    "transactionIndex": 1,
+    "gasUsed": "160624",
+    "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+    "blockHash": "0x146b9a0c47c57dfab4cd30e3242952aaeb3cbc5903b37c10282b48a6add85db3",
+    "transactionHash": "0x3572a3eef62da1b223a67567e89066dcd52b3857c7e34c188367f5701c0d08a3",
+    "logs": [],
+    "blockNumber": 9501358,
+    "cumulativeGasUsed": "160624",
+    "status": 1,
+    "byzantium": true
+  },
+  "args": [],
+  "numDeployments": 1,
+  "solcInputHash": "292f6ba2dbabdd1f5b177ec6945557f2",
+  "metadata": "{\"compiler\":{\"version\":\"0.8.15+commit.e14f2714\"},\"language\":\"Solidity\",\"output\":{\"abi\":[{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"message\",\"type\":\"bytes32\"},{\"components\":[{\"internalType\":\"uint8\",\"name\":\"v\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"r\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"s\",\"type\":\"bytes32\"}],\"internalType\":\"struct SignatureLib.Signature\",\"name\":\"signature\",\"type\":\"tuple\"}],\"name\":\"recoverSigner\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"pure\",\"type\":\"function\"}],\"devdoc\":{\"kind\":\"dev\",\"methods\":{},\"version\":1},\"userdoc\":{\"kind\":\"user\",\"methods\":{},\"version\":1}},\"settings\":{\"compilationTarget\":{\"contracts/SignatureLib.sol\":\"SignatureLib\"},\"evmVersion\":\"london\",\"libraries\":{},\"metadata\":{\"bytecodeHash\":\"ipfs\",\"useLiteralContent\":true},\"optimizer\":{\"enabled\":true,\"runs\":200},\"remappings\":[]},\"sources\":{\"contracts/SignatureLib.sol\":{\"content\":\"// SPDX-License-Identifier: MIT\\npragma solidity =0.8.15;\\n\\nlibrary SignatureLib {\\n    struct Signature {\\n        uint8 v;\\n        bytes32 r;\\n        bytes32 s;\\n    }\\n\\n    function recoverSigner(\\n        bytes32 message,\\n        Signature memory signature\\n    ) public pure returns (address) {\\n        bytes32 prefixedHash = keccak256(\\n            abi.encodePacked(\\\"\\\\x19Ethereum Signed Message:\\\\n32\\\", message)\\n        );\\n        return ecrecover(prefixedHash, signature.v, signature.r, signature.s);\\n    }\\n}\\n\",\"keccak256\":\"0x3d2aea16b52e98b22e10addc4401c67151de809a94688cc5fa7c0e5529b65ccb\",\"license\":\"MIT\"}},\"version\":1}",
+  "bytecode": "0x6101f061003a600b82828239805160001a60731461002d57634e487b7160e01b600052600060045260246000fd5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600436106100355760003560e01c80635cff288d1461003a575b600080fd5b61004d610048366004610126565b610069565b6040516001600160a01b03909116815260200160405180910390f35b6040517f19457468657265756d205369676e6564204d6573736167653a0a3332000000006020820152603c81018390526000908190605c0160408051601f198184030181528282528051602091820120865187830151888501516000875293860180865283905260ff9091169385019390935260608401929092526080830152915060019060a0016020604051602081039080840390855afa158015610113573d6000803e3d6000fd5b5050604051601f19015195945050505050565b600080828403608081121561013a57600080fd5b833592506060601f198201121561015057600080fd5b506040516060810181811067ffffffffffffffff8211171561018257634e487b7160e01b600052604160045260246000fd5b604052602084013560ff8116811461019957600080fd5b8152604084810135602083015260609094013593810193909352509290915056fea2646970667358221220cdc4ecf731940bf36c38f6a0b573693da49179cf315965269bbd00ea1ba2406664736f6c634300080f0033",
+  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600436106100355760003560e01c80635cff288d1461003a575b600080fd5b61004d610048366004610126565b610069565b6040516001600160a01b03909116815260200160405180910390f35b6040517f19457468657265756d205369676e6564204d6573736167653a0a3332000000006020820152603c81018390526000908190605c0160408051601f198184030181528282528051602091820120865187830151888501516000875293860180865283905260ff9091169385019390935260608401929092526080830152915060019060a0016020604051602081039080840390855afa158015610113573d6000803e3d6000fd5b5050604051601f19015195945050505050565b600080828403608081121561013a57600080fd5b833592506060601f198201121561015057600080fd5b506040516060810181811067ffffffffffffffff8211171561018257634e487b7160e01b600052604160045260246000fd5b604052602084013560ff8116811461019957600080fd5b8152604084810135602083015260609094013593810193909352509290915056fea2646970667358221220cdc4ecf731940bf36c38f6a0b573693da49179cf315965269bbd00ea1ba2406664736f6c634300080f0033",
+  "devdoc": {
+    "kind": "dev",
+    "methods": {},
+    "version": 1
+  },
+  "userdoc": {
+    "kind": "user",
+    "methods": {},
+    "version": 1
+  },
+  "storageLayout": {
+    "storage": [],
+    "types": null
+  }
+}
