+++ package.json
@@ -8,7 +8,7 @@
     "verify:uat": "npx hardhat --network goerli etherscan-verify --solc-input",
     "deploy:local": "npx hardhat compile && npx hardhat deploy --network localhost --tags underlying,oracle,registryToken,vault,market --export ./exports/localhost.json",
     "deploy:uat": "npx hardhat clean && npx hardhat compile && npx hardhat deploy --network goerli --tags oracle,vault,market --export ./exports/goerli.json && npx hardhat --network goerli etherscan-verify",
-    "deploy:uat2": "npx hardhat clean && npx hardhat compile && npx hardhat deploy --network arbitrumGoerli --tags underlying,oracle,vault,market --export ./exports/goerli.json && npx hardhat --network goerli etherscan-verify",
+    "deploy:uat2": "npx hardhat clean && npx hardhat compile && npx hardhat deploy --network arbitrumGoerli --tags underlying,oracle,vault,market --export ./exports/arbitrumGoerli.json && npx hardhat --network arbitrumGoerli etherscan-verify",
 
     "hh:node": "npx hardhat node --hostname localhost",
     "test": "npx hardhat test",
